fixing file errors in import

+++ src/controllers/admin/users-admin-controller/deleteUser.js
@@ -1,14 +1,24 @@
+import mongoose from "mongoose";
 import User from "../../../models/user-model.js";
 
 const deleteUser = async (req, res) => {
   try {
     const { userId } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
+
     const deletedUser = await User.findOneAndDelete({
-      _id: userId,
+      _id: new mongoose.Types.ObjectId(userId),
       status: "user",
     });
-    if (!deletedUser) throw new Error("Error while deleting");
+
+    if (!deletedUser) {
+      return res
+        .status(404)
+        .json({ message: "User not found or already deleted" });
+    }
 
     res.status(200).json({ message: "The user has been deleted" });
   } catch (error) {
